Revert "Translated public/src/client/topic/events.js to Typescript"
